Adds support for multiple meshes.
Previous method only worked with a single mesh because of the way the
single VAO interacted with the multiple VBOs. We now have a VAO per
mesh, but there's a lot of duplication and I'm sure there's a better
way. But the laptop just started shutting down at random intervals, so
no time to look.