Adds an Entity Component System-like structure
Now we're finally working with something like a typically ECS. Not sure
how far it will go or how much it will change (are we really okay with
entities only being allowed one of each component type? Are the
components even being used "correctly"?" but it's here and it sorta
works, though overhead probably could be reduced by a decent amount in a
number of place. But premature optimisation is the devil etc.