Full data-oriented ECS
The big one. We've got an ECS, nicely(?) templated and with easily
creatable components that can be attached and removed very simply.

Well, the interface is nice. Removal still isn't in and the entities
don't get destroyed so this code is all in all pretty unsafe. But it's a
good start.