Adds GetBitmask to Entity Manager
Adds a function to get a bitmask for a set of components from the
entity manager, and a Has function for an entity that takes in the
mask. This allows for only having to construct a single tuple for the
desired check, making a Has check on a list of entities a simple bitwise
operation and nothing more.