Adds RenderManager2D depth sorting
Sprites sent to the 2D renderer can now send a depth value, which will
change the batch in which it is sent, allowing for ordering draw calls.

The way this is done results in storing a vector within a map within a
map. Doesn't sound too great, but I can't know it's horribly detrimental
without benchmarking it. So it stays for now.