Fixes input detection for non-working keys
The Set functions for detecting input would bail if the value of the key
(SDL's value coming in) was greater than the size of the bitset holding
J_ values. This was a mistake accidentally put in from the correct check
on the Get functions. Also changes the way buttons were being retrieved,
and makes some of the gamepad checking a bit nicer.