Adds template HTML page with logo and loading bar
An outline of an HTML page has been created where all Insolence elements
(sans CSS include) are put in a div that can easily be copy and pasted
out to be integrated into another page..

The page shows progress with a loading bar and MB downloaded counter
with an XMLHttpRequest that looks for a .mem file - so compiling
as a JS project with the --memory-init-file flag enabled. All asset
dependency sizes are received through a Module function.