Fixes entity Has to work without default ctor
The Has function when taking in more than one component type (and ending
up in a place where it needs to be part of a tuple) wouldn't work with
components that didn't have a default constructor with no parameters
since the tuple was constructing them. The Has function now does all
checks against pointer types instead.