Adds static Game class for globals
Some things make sense being global - namely the entity manager and the
window - so they're put a singleton that is essentially in control of
the game state, of which there is now definitively a single instance.

GameWorld becomes GameLoop, but still handles most of the same stuff,
other than the creation of globals that now belong to game.