Adds newline support to TextRenderable
TextRenderable now supports \n for a new line and rewrites some of the
wrapping (including getting rid of allocation of a new string for a
different wrapped version).

The wrapping is still a little bit janky when the first word exceeds the
wrap length, but it's not a big deal for the current use case.